dma_mem: support >32-bit physical address space.
These changes follow an audit to make USDPAA safe for physical addresses
above 32-bits (4GB). It is yet to be tested on reserved memory above 4GB
though, so this work is not necessarily complete.

+++ drivers/dma_mem/setup.c
@@ -79,7 +79,7 @@ int dma_mem_setup(void)
 				DMA_MEM_SIZE - DMA_MEM_BPOOL);
 	if (ret)
 		goto err;
-	__dma_virt = (dma_addr_t)virt;
+	__dma_virt = (dma_addr_t)(unsigned long)virt;
 	printf("FSL dma_mem device mapped (phys=0x%x,virt=%p,sz=0x%x)\n",
 		DMA_MEM_PHYS, virt, DMA_MEM_SIZE);
 	return 0;
